requirements.txt
Django==2.1.3
djangorestframework==3.8.0

PyMySQL~=1.0.3
PyJWT==v1.7.1
coreapi~=2.3.3
msrest~=0.7.1
Pillow~=9.5.0
#psycopg2==2.8.6
azure-storage-blob==12.16.0
PyYAML==6.0

azure-cognitiveservices-vision-computervision==0.9.0




